+++ CHANGES.rst
@@ -1,3 +1,15 @@
+0.2.2
+-----
+-  Use `utils.get_workspace()` to avoid duplication
+
+0.2.2
+-----
+-  Fix development.ini file
+
+0.2.1
+-----
+-  Bump required version for praekelt-python-gitmodel
+
 0.2
 ---
 -  Added `git.content_repo_url` for cloning when app starts